small bug

+++ NeuralNetwork_with_CUDA/losses.cpp
@@ -34,7 +34,7 @@ float mae_loss(int** Y_true, float** Y_pred, int Y_rows, int Y_cols)
     for (int i = 0; i < Y_rows; i++) {
         temp = 0.0f;
         for (int j = 0; j < Y_cols; j++) {
-            temp += abs((Y_true[i][j]-(Y_pred[i][j])));
+            temp += abs((int)(Y_true[i][j]-(Y_pred[i][j])));
         }
         loss += temp / Y_rows;
     }
